clean up : cleaning up code and optimize unused codes
