+++ src/main/java/model/Deadline.java
@@ -1,7 +1,7 @@
-package Model;
+package model;
 
 public class Deadline extends Task {
-	private String deadline;
+	private final String deadline;
 	
 	public Deadline(String desc, String deadline) {
 		super(desc);
