+++ src/main/java/model/ToDos.java
@@ -1,4 +1,4 @@
-package Model;
+package model;
 
 public class ToDos extends Task {
 	public ToDos(String desc) {
